enhanced FAQ page to match design
