+++ src/assets/data/FAQ.jsx
@@ -1,0 +1,72 @@
+// FAQ Questions and Answers in a list
+const info = [
+    {
+      question: 'How do I create an account?',
+      answer:
+        "Yay! We’re excited that you want to try out our services. To create an account, please click the “create an account” button on the upper right corner of the first page you see when you visit our website. This button takes you to another page, where you will be asked to fill in some of your details. You can also create an account by signing up with your Google and Facebook account, if you already have one. Please ensure you supply accurate details so that you do not encounter any difficulties while creating your account.",
+    },
+    {
+      question: 'How do I contact customer care?',
+      answer:
+        'You can contact customer care by calling.........  For less urgent issues, you can chat up the Admin through the admin chat box or you can fill out a form in the .......',
+    },
+    {
+      question: 'What type of product can i sell on this platform?',
+      answer:
+        'This platform enables users to buy and sell tech gadgets like laptops, phones, routers, televisions, headphones, drives, monitors and other similar items. Please click here to see the full list of products we allow on this platform.',
+    },
+    {
+      question: 'How can I report fraudulent activities',
+      answer:
+        'If you suspect or confirm any fraudulent activity by a buyer or seller on this platform, please do not hesitate to contact the Admin via the admin chat box on your dashboard. You can also make a formal report by filling out a form in the help and support feature on your dashboard.  For more urgent cases, please our call customer service care at ............, an operator will promptly attend to you. ',
+    },
+    {
+        question: 'How do I pay for items?',
+        answer:
+          'This platform does not carry out payment transactions. To pay for an item, you will have to agree with the seller on the mode, time and place of payment.',
+      },
+    {
+        question: 'How can I swap my products with other sellers?',
+        answer:
+          'To swap your product with other sellers’, kindly reach out to the seller you want to swap with. You can swap your products with another seller when you both have negotiated and agreed on the terms of the swap.',
+      },
+    {
+        question: 'How do I protect myself from being scammed?',
+        answer:
+          'Techmart will not be held responsible in the event that you are scammed while making use of our platform. Please check out numbers 9 and 10 of our Terms and Conditions here for more information. To prevent yourself from being scammed, please be vigilant and careful in your dealings with a potential buyer or seller. Ensure that you see the item you intend to purchase and inspect it thoroughly before making any payment.',
+      },
+      {
+        question: 'Can I edit my ad after posting it?',
+        answer:
+          'Yes you can edit an ad after posting. To edit your add click on......',
+      },
+      {
+        question: 'Why was my product ad declined?',
+        answer:
+          'It is possible you uploaded a product that we do not allow on the platform. Please check here to see the list of products we allow on TechMart.',
+      },
+      {
+        question: 'I bought a fake product. What do I do?',
+        answer:
+          'Our non interference policy does not allow us interfere in issues involving buyers and sellers on this platform. Kindly refer to numbers 9 and 10 of our Terms and Conditions for more clarity. Please do well promptly report the account of the seller who sold you a fake product so we can make our own ivestigations and prevent other buyers from falling victims. ',
+      },
+      {
+        question: 'Am I entitled to home delivery after purchasing a product?',
+        answer:
+          'Unfortunately, home delivery is not part of the services we provide. If you want your purchased items to be delivered to your home, kindly reach an agreement to that effect with the seller you purcased the items from.',
+      },
+      {
+        question: 'Whose duty it is to inspect the products?',
+        answer:
+          'It is your duty as a buyer to inspect the products you purchase. If you agree with the seller to inspect or pay for a product or item physically, please be safety conscious.',
+      },
+      {
+        question: 'How can I delete my account?',
+        answer:
+          '...',
+      },
+
+  ];
+  
+
+  export default info;
